move x509 utils to common package
