+++ gateway-service/src/main/java/org/zowe/apiml/gateway/filters/AbstractAuthSchemeFactory.java
@@ -24,8 +24,8 @@ import org.springframework.web.reactive.function.client.ClientResponse;
 import org.springframework.web.server.ServerWebExchange;
 import org.zowe.apiml.constants.ApimlConstants;
 import org.zowe.apiml.gateway.service.InstanceInfoService;
-import org.zowe.apiml.gateway.x509.X509Util;
 import org.zowe.apiml.message.core.MessageService;
+import org.zowe.apiml.security.common.util.X509Util;
 import org.zowe.apiml.util.CookieUtil;
 import reactor.core.publisher.Mono;
